ingestion/requirements.txt
qdrant-client>=1.7.0
pypdf>=3.17.0
python-docx>=1.1.0
tiktoken>=0.5.0
requests>=2.31.0
watchdog>=3.0.0
